Update vendored mupdf 1.27.2 -> 1.28.0-rc2
Final step of updating mupdf and its dependencies to 1.28.0-rc2.

Per-file 3-way merge (base 1.27.2 / vendored / 1.28.0-rc2) preserving all
SumatraPDF local edits (marked and unmarked, e.g. fz_set_optind, the
tools/*_to_buffer entry points). 6 conflicts resolved: filter-dct, pdf-op-run
(#5285 t3 fix now upstream-native), stext-search, html-parse -> upstream;
pdf-link keeps its marker; epub-doc -> upstream and drops the now-redundant
chapters_cached cache (upstream changed spine to an indexed array + spine_len).

New compiled sources added to premake (cull-device.c, options.c,
pdf-struct.c). Markdown (source/html/md.c) is disabled (FZ_ENABLE_MD=0) and
not compiled because it needs the cmark-gfm thirdparty lib, which we do not
vendor; md_document_handler is therefore not registered. Disabled benign
C4146 in the mupdf project.

+++ mupdf/source/fitz/filter-fax.c
@@ -370,7 +370,7 @@ fill_bits(fz_context *ctx, fz_faxd *fax)
 		if (c == EOF)
 			return EOF;
 		fax->bidx -= 8;
-		fax->word |= c << fax->bidx;
+		fax->word |= (unsigned)c << fax->bidx;
 	}
 	return 0;
 }
